requirements.txt
streamlit>=1.35.0
openpyxl>=3.1.2
PyMuPDF>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.10
pypdf>=4.2.0
deep-translator>=1.11.4
